requirements.txt
tkinter
google-generativeai
